Added login test and implemented mark.incremental but not working perfectly as intended
