+++ ui-automation/.gitignore
@@ -1,3 +1,1 @@
-chromedriver.exe
-geckodriver.exe
-settings.py
+basictests/
