--- ui-automation/readme.md
@@ -1,1 +1,0 @@
-add drivers to this folder but remember to add to gitignore
