+++ ui-automation/test_runs/.gitignore
@@ -1,0 +1,3 @@
+chromedriver.exe
+geckodriver.exe
+settings.py
