+++ ui-automation/test_runs/conftest.py
@@ -1,0 +1,24 @@
+import pytest
+from selenium import webdriver
+
+def pytest_runtest_makereport(item, call):
+    if "incremental" in item.keywords:
+        if call.excinfo is not None:
+            parent = item.parent
+            parent._previousfailed = item
+
+def pytest_runtest_setup(item):
+    if "incremental" in item.keywords:
+        previousfailed = getattr(item.parent, "_previousfailed", None)
+        if previousfailed is not None:
+            pytest.xfail("requisite test failed (%s)" %previousfailed.name)
+
+@pytest.fixture(params=["chrome", "firefox"], scope="class")
+def driver_init(request):
+    if request.param == "chrome":
+        web_driver = webdriver.Chrome()
+    if request.param == "firefox":
+        web_driver = webdriver.Firefox()
+    request.cls.driver = web_driver
+    yield
+    web_driver.close()
